requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-ckeditor==6.0.0
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.11.0
django-redis==4.12.1
fabric==2.5.0
Faker==4.4.0
gunicorn==20.0.4
invoke==1.4.1
paramiko==2.7.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
